blockchain_layer/services/data_storage/requirements.txt
fastapi==0.95.1
uvicorn
pydantic
requests==2.28.2
Jinja2==3.1.2
debugpy # Required for debugging.
google-cloud-firestore==2.11.1
python-dotenv==0.21.0